Issue #17 - Move constants and miscellaneous functions from utils to separate files
