+++ src/js/postHandler.js
@@ -5,9 +5,11 @@ import _ from "lodash";
 import Mustache from "mustache";
 
 import {
-  StoreKey,
   LocalStore
 } from "./utils";
+import {
+  StoreKey,
+} from "./shared/constants";
 
 // Wait for the page to be loaded to execute the initialization function.
 document.addEventListener("DOMContentLoaded", function() {
