+++ src/js/utils.js
@@ -3,130 +3,7 @@ import Mustache from "mustache";
 import Notiflix from "notiflix";
 import { Sortable } from "sortablejs";
 
-// --- Constants --- //
-
-export var MiscURLs = {
-  ABOUT_US_URL: "https://www.criticalstart.com/company/",
-  EXTENSION_HOME_URL: "http://www.criticalstart.com/threat-analytics-chrome-plugin/",
-  INSTALLED_URL: "https://www.criticalstart.com/threat-analytics-chrome-plugin/",
-  SUPPORT_EMAIL: "support@criticalstart.com",
-};
-
-export var BasicConfig = {
-  CONFIG_URL: 0,
-  USE_GROUPS: 1,
-  ENCRYPTED: 2,
-  ENCRIPTION_KEY: 3,
-  AUTO_UPDATE: 4,
-};
-
-export var SearchProv = {
-  MENU_INDEX: 0,
-  LABEL: 1,
-  LINK: 2,
-  ENABLED: 3,
-  FROM_CONFIG: 4,
-  GROUP: 5,
-  IS_POST: 6,
-  POST_REQUEST: 7,
-  PROXY_ENABLED: 8,
-  PROXY_URL: 9
-};
-
-export var ProvGroups = {
-  NUMBER: 0,
-  NAME: 1,
-};
-
-export var ProvQuery = {
-  MENU_INDEX: 0,
-  LABEL: 1,
-  QUERY: 2,
-  ENABLED: 3,
-};
-
-export var StoreKey = {
-  CARBON_BLACK: "carbon_black",
-  LAST_CONFIG_DATA: "last_config_data",
-  NET_WITNESS: "netwitness_investigator",
-  RSA_SECURITY: "rsa_security_analytics",
-  SEARCH_PROVIDERS: "search_providers",
-  SETTINGS: "settings",
-};
-
-export var CBC_CONFIG = [
-  { key: "CBCConfigEnable", type: "checkbox", label: "Enable Carbon Black Settings" },
-  { key: "CBCConfigUseHttps", type: "checkbox", label: "HTTPS (SSL) Enabled" },
-  { key: "CBCConfigHost", type: "text", label: "Carbon Black Hostname or IP Address" },
-  { key: "CBCConfigPort", type: "text", label: "Port Number (Leave blank if port 80 and HTTP or port 443 and HTTPS)" },
-  { key: "CBCConfigURLVersion", type: "text", label: "Carbon Black URL Version (Default is 1)" },
-  { key: "CBCConfigNewTab", type: "checkbox", label: "Switch Focus to New Tab" },
-  { key: "CBCConfigPopup", type: "checkbox", label: "Enable debug popup window" },
-];
-
-export var NWI_CONFIG = [
-  { key: "NWIConfigEnable", type: "checkbox", label: "Enable NetWitness Investigator Settings" },
-  { key: "NWIConfigExampleLink", type: "text", label: "Paste example link from NetWitness Investigator to autofill settings", autofiller: true, placeholder: "e.g. https://netwitness.com:81/?collection=12" },
-  { key: "NWIConfigHost", type: "text", label: "Host (IP Address/Hostname)" },
-  { key: "NWIConfigPort", type: "text", label: "Port Number" },
-  { key: "NWIConfigCollectionName", type: "text", label: "Collection Name" },
-  { key: "NWIConfigGMT", type: "checkbox", label: "Use GMT time (no time zone adjustment)" },
-  { key: "NWIConfigRange1", type: "number", label: "Search Range 1 in Hour(s)" },
-  { key: "NWIConfigRange2", type: "number", label: "Search Range 2 in Hour(s)" },
-  { key: "NWIConfigRange3", type: "number", label: "Search Range 3 in Hour(s)" },
-  { key: "NWIConfigRange4", type: "number", label: "Search Range 4 in Hour(s)" },
-  { key: "NWIConfigPopup", type: "checkbox", label: "Enable debug popup window" },
-];
-
-export var NWI_RANGE_LENGTH = 4;
-
-export var RSA_CONFIG = [
-  { key: "RSAConfigEnable", type: "checkbox", label: "Enable RSA Security Analytics Settings" },
-  { key: "RSAConfigExampleLink", type: "text", label: "Paste example link from Security Analytics Investigation to autofill settings", autofiller: true, placeholder: "e.g. https://security.com:81/investigation/12/" },
-  { key: "RSAConfigHost", type: "text", label: "Security Analytics Host (IP Address/Hostname)" },
-  { key: "RSAConfigPort", type: "text", label: "Port Number (leave blank if port 80 and HTTP or port 443 and HTTPS)" },
-  { key: "RSAConfigDevId", type: "text", label: "Device ID" },
-  { key: "RSAConfigUseHttps", type: "checkbox", label: "HTTPS (SSL) Enabled" },
-  { key: "RSAConfigRange1", type: "number", label: "Search Range 1 in Hour(s)" },
-  { key: "RSAConfigRange2", type: "number", label: "Search Range 2 in Hour(s)" },
-  { key: "RSAConfigRange3", type: "number", label: "Search Range 3 in Hour(s)" },
-  { key: "RSAConfigRange4", type: "number", label: "Search Range 4 in Hour(s)" },
-  { key: "RSAConfigNewTab", type: "checkbox", label: "Switch Focus to New Tab" },
-  { key: "RSAConfigPopup", type: "checkbox", label: "Enable debug popup window" },
-];
-
-export var RSA_RANGE_LENGTH = 4;
-
-
-// --- Miscellaneous --- //
-
-export function getGroupProviders(groupIndex, providers) {
-  var mask = Math.pow(2, groupIndex)
-  return _.filter(providers, function(item) {
-    // Note that the group's value (from providers) uses each bit of the number for indicate to
-    // which groups it belongs; so "1" (0b0001) it's for group 1, "2" (0b0010) it's for group 2,
-    // "3" (0b0011) it's for groups 1 and 2, and "4" (0b0100) it's for group 3.
-    return item.enabled && (item.group & mask);
-  });
-}
-
-export function getProviderTargetURL(provider, selectionText) {
-  var targetURL;
-
-  if (provider.postEnabled === true || provider.postEnabled === 'true') {
-    targetURL = "postHandler.html?name=" + provider.label + "&data=" + selectionText;
-  } else {
-    targetURL = provider.link.replace(/TESTSEARCH/g, selectionText);
-    targetURL = targetURL.replace(/%s/g, selectionText);
-    targetURL = targetURL.replace(/TESTB64SEARCH/g, encodeURIComponent(btoa(selectionText)));    
-  }
-
-  return targetURL;
-}
-
-export var isDate = function (date) {
-  return new Date(date) !== "Invalid Date" && !isNaN(new Date(date));
-};
+import { StoreKey } from "./shared/constants";
 
 // --- Storage --- //
 
@@ -768,15 +645,6 @@ export function providerTabHelper(
 };
 
 export default {
-  MiscURLs,
-  StoreKey,
-  CBC_CONFIG,
-  NWI_CONFIG,
-  NWI_RANGE_LENGTH,
-  RSA_CONFIG,
-  RSA_RANGE_LENGTH,
-  getGroupProviders,
-  getProviderTargetURL,
   Storage,
   LocalStore,
   ConfigFile,
